Fixed NPE in formentry module when duplicating the basic form

+++ src/org/openmrs/module/formentry/advice/DuplicateFormAdvisor.java
@@ -74,7 +74,8 @@ public class DuplicateFormAdvisor extends StaticMethodMatcherPointcutAdvisor imp
 			}
 			finally {
 				try {
-					OpenmrsUtil.deleteDirectory(tempDir);
+					if (tempDir != null)
+						OpenmrsUtil.deleteDirectory(tempDir);
 				} 
 				catch (IOException ioe) {
 					// pass
